lesson_10/requirements.txt
selenium==4.21.0
pytest==8.2.1
pytest-allure-adaptor==1.7.10
allure-pytest==2.13.5
allure-python-commons==2.13.5
webdriver-manager==4.0.1